Merge query string version into master (#1)
Currently, this extension does not redirect if a query string is present.  If a link is clicked from their email newsletter, which contains utm_* campaign information, this extension will not redirect.

The changes submitted here extend the URL patterns that this extension applies, such that a URL with a query string is accepted.  Furthermore, a check is now performed that evaluates the presence of a question mark (indicating the beginning of the query string), and either adds "?ALLSTEPS" or "&ALLSTEPS" depending on if a query string is already present.

* Change version to 0.2
* Change URL match pattern in manifest
* Change URL match pattern in background.js
* Add extra conditional in background.js checking if a query string is already present